# server/cmd/docs/swagger.yaml
definitions:
  request.Anchor:
    properties:
      id:
        type: string
      x:
        type: number
      "y":
        type: number
    required:
    - id
    - x
    - "y"
    type: object
  request.AnchorsRequest:
    properties:
      anchors:
        items:
          $ref: '#/definitions/request.Anchor'
        type: array
    required:
    - anchors
    type: object
...
  response.Anchor:
    properties:
      id:
        type: string
      x:
        type: number
      "y":
        type: number
    required:
    - id
    - x
    - "y"
    type: object
...
info:
  contact: {}
  title: Lynx API
  version: "1.0"
paths:
...
  /anchors:
    get:
      operationId: getAllAnchors
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.Anchor'
            type: array
      summary: GetAllAnchors
      tags:
      - anchor
    post:
      operationId: upsertAnchors
      parameters:
      - description: Anchors Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AnchorsRequest'
      responses:
        "201":
          description: Created
      summary: UpsertAnchors
      tags:
      - anchor
...
swagger: "2.0"