# server/internal/infrastructure/postgres/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package sqlc

import (
	"time"

	"github.com/google/uuid"
)

type Anchor struct {
	ID        uuid.UUID
	AnchorID  string
	X         float64
	Y         float64
	CreatedAt time.Time
	UpdatedAt time.Time
}

type Device struct {
	ID        uuid.UUID
	DeviceID  string
	Status    string
	X         float64
	Y         float64
	CreatedAt time.Time
	UpdatedAt time.Time
}

type Scale struct {
	ID        uuid.UUID
	Meters    float64
	Pixels    float64
	CreatedAt time.Time
	UpdatedAt time.Time
}

type User struct {
	ID           int64
	BetterAuthID string
	CreatedAt    time.Time
	UpdatedAt    time.Time
}